refactor: 기능별 패키지 리펙토링
