+++ backend/src/main/java/RunningMate/backend/domain/running/entity/GroupTag.java
@@ -1,4 +1,4 @@
-package RunningMate.backend.DAO;
+package RunningMate.backend.domain.running.entity;
 
 public enum GroupTag {
     BEGINNER, INTERMEDIATE, EXPERT, ATHLETE
